Fixed demo.js and added new docs
This should allow you to use demo.js as a place to experiment with HTML element manipulation and have variables visible directly in the browser console.

The reason why demo.js wasn't working yesterday was because the async attribute was missing from the script tag.

Without the async tag, the query selectors try to modify the DOM objects when they haven't been loaded yet. My understanding is that the async tag allows the script to load after the DOM objects have been fully loaded.